requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
psycopg2-binary
boto3